[refactor] Add directory, move config files(inc/config, src/config)
