+++ kangkim_Makefile
@@ -1,11 +1,11 @@
 CXX			= c++
 # CXXFLAGS	= -std=c++98 -Wall -Wextra -Werror
 CXXFLAGS	= -std=c++11 -Wall -Wextra -Werror
-INC			= inc
+INC			= inc/config/
 
 CONFIG_TEST_NAME	= config_test
 CONFIG_SRC	= Config.cpp Directive.cpp LocationBlock.cpp ServerBlock.cpp 
-SRCDIR		= src/
+SRCDIR		= src/config/
 
 CONFIG_SRCS	:= $(addprefix $(SRCDIR), $(CONFIG_SRC)) test/config/main.cpp
 CONFIG_OBJS	= $(CONFIG_SRCS:.cpp=.o)
